test: split runtime status resume tests
